+++ Ribbon/Properties/AssemblyInfo.cs
@@ -37,5 +37,5 @@ using System.Runtime.InteropServices;
 #else
 [assembly: AssemblyVersion("1.0.0.0")]
 #endif
-[assembly: AssemblyFileVersion("2.14.0.0")]
+[assembly: AssemblyFileVersion("2.15.0.0")]